requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.3
black==26.1.0
blinker==1.9.0
boto3==1.35.10
botocore==1.35.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
greenlet==3.3.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jq==1.11.0
librt==0.7.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyMySQL==1.1.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.4.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.46
starlette==0.37.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.2.2
uvicorn==0.25.0
Werkzeug==3.1.6
